Finished carousel logic | Moved footer & header to general folder
