+++ src/assets/styles/components/carousel.scss
@@ -1,0 +1,15 @@
+.carousel-container{
+    padding: 2rem;
+    width: 100%;
+
+    display: flex;
+    justify-content: flex-start;
+    align-items: center;
+    flex-wrap: nowrap;
+    overflow-x: scroll;
+
+    .carousel-card{
+        flex-shrink: 0;
+        margin: 0 .8rem;
+    }
+}
