+++ src/components/general/Carousel.tsx
@@ -1,8 +1,9 @@
+import '../../assets/styles/components/carousel.scss';
 import React, { FC } from 'react';
 
 const CarouselComponent: FC= ({children}) => {
     return (
-        <div>
+        <div className="carousel-container">
             {children}
         </div>
     )
