+++ src/components/general/Footer.tsx
@@ -1,7 +1,7 @@
 import '../../assets/styles/components/footer.scss';
-import React from 'react'
+import React from 'react';
 import { mockNavFooterItems } from '../../utils/constants/navData';
-import NavLinks from '../general/NavLinks';
+import NavLinks from './NavLinks';
 
 const FooterComponent = () => {
     return (
