+++ src/navigation/navigation.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { BrowserRouter, Redirect, Route, Switch } from "react-router-dom";
-import FooterComponent from '../components/footer/Footer';
-import Header from '../components/header/Header';
+import FooterComponent from '../components/general/Footer';
+import Header from '../components/general/Header';
 import HomePageComponent from '../features/HomePage';
 
 const NavigationComponent = () => {
